@@ -9,14 +9,14 @@ homepage: https://github.com/ayalma/feature_discovery
 environment:
   # This should match the Flutter stable SDK constraint.
   # See https://github.com/flutter/flutter/blob/stable/packages/flutter/pubspec.yaml.
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.0.4
-  shared_preferences: ^0.5.6+1
+  provider: ^5.0.0
+  shared_preferences: ^2.0.3
 
 dev_dependencies:
   flutter_test:
