@@ -103,30 +103,32 @@ class _MyHomePageState extends State<MyHomePage> {
               children: <Widget>[
                 const Text(
                     'Also, notice how the pulsing animation is not playing because it is deactivated for this feature.'),
-                FlatButton(
-                    child: Text('Toggle enablePulsingAnimation',
-                        style: Theme.of(context)
-                            .textTheme
-                            .button
-                            .copyWith(color: Colors.white)),
-                    onPressed: () => setState(() {
-                          feature1EnablePulsingAnimation =
-                              !feature1EnablePulsingAnimation;
-                        })),
+                TextButton(
+                  onPressed: () => setState(() {
+                    feature1EnablePulsingAnimation =
+                        !feature1EnablePulsingAnimation;
+                  }),
+                  child: Text('Toggle enablePulsingAnimation',
+                      style: Theme.of(context)
+                          .textTheme
+                          .button
+                          .copyWith(color: Colors.white)),
+                ),
                 const Text(
                     'Ignore the items below or tap the button to toggle between OverflowMode.clip and OverflowMode.doNothing!'),
-                FlatButton(
-                    child: Text('Toggle overflowMode',
-                        style: Theme.of(context)
-                            .textTheme
-                            .button
-                            .copyWith(color: Colors.white)),
-                    onPressed: () => setState(() {
-                          feature1OverflowMode =
-                              feature1OverflowMode == OverflowMode.clipContent
-                                  ? OverflowMode.ignore
-                                  : OverflowMode.clipContent;
-                        })),
+                TextButton(
+                  onPressed: () => setState(() {
+                    feature1OverflowMode =
+                        feature1OverflowMode == OverflowMode.clipContent
+                            ? OverflowMode.ignore
+                            : OverflowMode.clipContent;
+                  }),
+                  child: Text('Toggle overflowMode',
+                      style: Theme.of(context)
+                          .textTheme
+                          .button
+                          .copyWith(color: Colors.white)),
+                ),
                 for (int n = 42; n > 0; n--)
                   const Text('Testing clipping (ignore or toggle)',
                       style: TextStyle(backgroundColor: Colors.black)),
@@ -149,24 +151,26 @@ class _MyHomePageState extends State<MyHomePage> {
               children: <Widget>[
                 const Text(
                     'Tap the magnifying glass to quickly scan your compounds'),
-                FlatButton(
-                  padding: const EdgeInsets.all(0),
-                  child: Text('Understood',
-                      style: Theme.of(context)
-                          .textTheme
-                          .button
-                          .copyWith(color: Colors.white)),
+                TextButton(
                   onPressed: () async =>
                       FeatureDiscovery.completeCurrentStep(context),
+                  child: Text(
+                    'Understood',
+                    style: Theme.of(context)
+                        .textTheme
+                        .button
+                        .copyWith(color: Colors.white),
+                  ),
                 ),
-                FlatButton(
-                  padding: const EdgeInsets.all(0),
-                  child: Text('Dismiss',
-                      style: Theme.of(context)
-                          .textTheme
-                          .button
-                          .copyWith(color: Colors.white)),
+                TextButton(
                   onPressed: () => FeatureDiscovery.dismissAll(context),
+                  child: Text(
+                    'Dismiss',
+                    style: Theme.of(context)
+                        .textTheme
+                        .button
+                        .copyWith(color: Colors.white),
+                  ),
                 ),
               ],
             ),
@@ -190,15 +194,16 @@ class _MyHomePageState extends State<MyHomePage> {
           description: Column(children: <Widget>[
             const Text(
                 'This is overly long to test OverflowMode.extendBackground. The green circle should be large enough to cover all of the text.'),
-            FlatButton(
-                child: Text('Add another item',
-                    style: Theme.of(context)
-                        .textTheme
-                        .button
-                        .copyWith(color: Colors.white)),
-                onPressed: () => setState(() {
-                      feature3ItemCount++;
-                    })),
+            TextButton(
+              onPressed: () => setState(() {
+                feature3ItemCount++;
+              }),
+              child: Text('Add another item',
+                  style: Theme.of(context)
+                      .textTheme
+                      .button
+                      .copyWith(color: Colors.white)),
+            ),
             for (int n = feature3ItemCount; n > 0; n--)
               const Text('Testing OverflowMode.extendBackground'),
           ]),
@@ -323,8 +328,7 @@ class _ContentState extends State<Content> {
                   contentLocation: ContentLocation.below,
                   child: EnsureVisible(
                     key: ensureKey,
-                    child: RaisedButton(
-                      child: const Text('Start Feature Discovery'),
+                    child: ElevatedButton(
                       onPressed: () {
                         FeatureDiscovery.discoverFeatures(
                           context,
@@ -338,6 +342,7 @@ class _ContentState extends State<Content> {
                           },
                         );
                       },
+                      child: const Text('Start Feature Discovery'),
                     ),
                   ),
                 ),
@@ -368,16 +373,16 @@ class _ContentState extends State<Content> {
                   description: Column(children: <Widget>[
                     const Text(
                         'You can test OverflowMode.wrapBackground here.'),
-                    FlatButton(
-                        padding: const EdgeInsets.all(0),
-                        child: Text('Add item',
-                            style: Theme.of(context)
-                                .textTheme
-                                .button
-                                .copyWith(color: Colors.white)),
-                        onPressed: () => setState(() {
-                              feature6ItemCount++;
-                            })),
+                    TextButton(
+                      onPressed: () => setState(() {
+                        feature6ItemCount++;
+                      }),
+                      child: Text('Add item',
+                          style: Theme.of(context)
+                              .textTheme
+                              .button
+                              .copyWith(color: Colors.white)),
+                    ),
                     for (int n = feature6ItemCount; n > 0; n--)
                       const Text('Testing OverflowMode.wrapBackground'),
                   ]),
@@ -416,10 +421,10 @@ class _ContentState extends State<Content> {
               child: FloatingActionButton(
                 backgroundColor: Colors.white,
                 foregroundColor: Colors.blue,
-                child: const Icon(Icons.drive_eta),
                 onPressed: () {
                   print('Floating action button tapped.');
                 },
+                child: const Icon(Icons.drive_eta),
               ),
             ),
           ),
