@@ -9,11 +9,11 @@ abstract class PersistenceProvider {
 
   /// Returns the list of steps (as strings) that the user has previously completed from
   /// the provided [featuresIds] set.
-  Future<Set<String>> completedSteps(Iterable<String> featuresIds);
+  Future<Set<String?>> completedSteps(Iterable<String?>? featuresIds);
 
   /// Informs the persistence layer that the user has completed the step identified by
   /// [featureId], and that it should record it in the persistence layer.
-  Future<void> completeStep(String featureId);
+  Future<void> completeStep(String? featureId);
 
   /// Requests that the persistence layer should remove the completion of the step
   /// identified by [featureId].
@@ -31,7 +31,7 @@ class SharedPreferencesProvider implements PersistenceProvider {
   /// If [sharedPrefsPrefix] is provided, it will be prepended to all step identifiers passed
   /// to the different methods. If [sharedPrefsPrefix] is not provided, the step
   /// identifiers will be used as-is.
-  const SharedPreferencesProvider([String sharedPrefsPrefix])
+  const SharedPreferencesProvider([String? sharedPrefsPrefix])
       : sharedPrefsPrefix = sharedPrefsPrefix ?? '';
 
   /// Use this string a prefix for all steps identifiers.
@@ -45,16 +45,16 @@ class SharedPreferencesProvider implements PersistenceProvider {
   }
 
   @override
-  Future<Set<String>> completedSteps(Iterable<String> featuresIds) async {
+  Future<Set<String?>> completedSteps(Iterable<String?>? featuresIds) async {
     final prefs = await SharedPreferences.getInstance();
-    return featuresIds
+    return featuresIds!
         .where((featureId) =>
             prefs.getBool(_normalizeFeatureId(featureId)) == true)
         .toSet();
   }
 
   @override
-  Future<void> completeStep(String featureId) async {
+  Future<void> completeStep(String? featureId) async {
     final prefs = await SharedPreferences.getInstance();
     await prefs.setBool(_normalizeFeatureId(featureId), true);
   }
@@ -74,7 +74,7 @@ class SharedPreferencesProvider implements PersistenceProvider {
     await Future.wait(mapResult);
   }
 
-  String _normalizeFeatureId(String featureId) =>
+  String _normalizeFeatureId(String? featureId) =>
       '$sharedPrefsPrefix$featureId';
 }
 
@@ -86,26 +86,24 @@ class SharedPreferencesProvider implements PersistenceProvider {
 /// This is a great implementation for testing.
 class MemoryPersistenceProvider implements PersistenceProvider {
   /// Instantiates a new [MemoryPersistenceProvider] with an initial completed set of steps.
-  const MemoryPersistenceProvider(Set<String> steps)
-      : assert(steps != null),
-        _steps = steps;
+  const MemoryPersistenceProvider(Set<String?> steps) : _steps = steps;
 
   /// Instantiates an empty [MemoryPersistenceProvider] instance.
   factory MemoryPersistenceProvider.empty() =>
-      const MemoryPersistenceProvider(<String>{});
+      const MemoryPersistenceProvider(<String?>{});
 
-  final Set<String> _steps;
+  final Set<String?> _steps;
 
   @override
   Future<bool> hasCompletedStep(String featureId) async =>
       _steps.contains(featureId);
 
   @override
-  Future<Set<String>> completedSteps(Iterable<String> featuresIds) async =>
-      featuresIds.where((featureId) => _steps.contains(featureId)).toSet();
+  Future<Set<String?>> completedSteps(Iterable<String?>? featuresIds) async =>
+      featuresIds!.where((featureId) => _steps.contains(featureId)).toSet();
 
   @override
-  Future<void> completeStep(String featureId) async => _steps.add(featureId);
+  Future<void> completeStep(String? featureId) async => _steps.add(featureId);
 
   @override
   Future<void> clearStep(String featureId) async => _steps.remove(featureId);
@@ -125,11 +123,11 @@ class NoPersistenceProvider implements PersistenceProvider {
   Future<bool> hasCompletedStep(String featureId) async => false;
 
   @override
-  Future<Set<String>> completedSteps(Iterable<String> featuresIds) async =>
-      <String>{};
+  Future<Set<String?>> completedSteps(Iterable<String?>? featuresIds) async =>
+      <String?>{};
 
   @override
-  Future<void> completeStep(String featureId) async {
+  Future<void> completeStep(String? featureId) async {
     // NO-OP
   }
 
